Refactor accounts-app
