+++ packages/accounts-app/src/Accounts/Accounts.tsx
@@ -2,13 +2,20 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import React from 'react';
+import React, { useContext } from 'react';
 import { Route, Switch } from 'react-router-dom';
 
+import { KeyringContext } from '../../../light-apps/src/context/KeyringContext';
 import { AddAccount } from './AddAccount';
 import { AccountsOverview } from './Overview';
 
 export function Accounts(): React.ReactElement {
+  const { accounts } = useContext(KeyringContext);
+
+  if (!Object.keys(accounts).length) {
+    return <Route path='/accounts/add' component={AddAccount} />;
+  }
+
   return (
     <Switch>
       <Route exact path='/accounts' component={AccountsOverview} />
