+++ packages/accounts-app/src/Accounts/index.ts
@@ -2,4 +2,4 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-export * from './AccountsOverview';
+export * from './Accounts';
