+++ packages/accounts-app/src/Addresses/index.ts
@@ -2,4 +2,4 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-export * from './ManageAddresses';
+export * from './Addresses';
