+++ packages/accounts-app/src/index.ts
@@ -3,6 +3,4 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 export * from './Accounts';
-export * from './AddAccount';
-export * from './ManageAddresses';
-export * from './Overview';
+export * from './Addresses';
